#85 updated comments on DropdownTogglePicker
